move User model namspace to App\Entities
