+++ database/factories/UserFactory.php
@@ -1,7 +1,7 @@
 <?php
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
-use App\User;
+use App\Entities\User;
 use Faker\Generator as Faker;
 use Illuminate\Support\Str;
